workflow: Add the spellcheck.yml workflow to itself!
Add a workflow that performs spellchecking on this repository. Also
clean up how the spellcheck.yaml and wordlist.txt files are defined
and passed into the docker container. Start an examples folder so we
can provide meaningful examples to users.

Note since we workflow off v0.6.0 of the action for now we have to
include a .spellcheck.yaml and a .wordlist.txt in the top-level. We
will remove them in a subsequent commit when we bump to 0.7.0.